refactor actor
